+++ CHANGES.rst
@@ -1839,6 +1839,10 @@ astropy.stats
 astropy.table
 ^^^^^^^^^^^^^
 
+- Fix ``.quantity`` property of ``Column`` class for function-units (e.g.,
+  ``dex``). Previously setting this was possible, but getting raised
+  an error. [#8425]
+
 astropy.tests
 ^^^^^^^^^^^^^